[MODIFY] move config/snackbar to utils/snackbar
